pyproject.toml
[build-system]
requires = ["hatchling"] # list of packages needed to build your package
build-backend = "hatchling.build" # name of the python object that frontends use to perform the build

[project]
name = "preplib"
version = "0.0.1"
authors = [
  { name="numlims", email="numlims@med.uni-greifswald.de"}
]
description = "help with ap8 csv preperation"
readme = "readme.md"
# requires-python = ">=3.8"
dependencies = [
  "pandas",
  "pyyaml",
  "traction @ git+https://github.com/numlims/traction"
]
classifiers = [
  "Programming Language :: Python :: 3"
]
